refactor(application/web): Controller 분리 - 데이터를 반환하는 RestController, View를 반환하는 Controller 분리
