+++ src/main/java/chess/application/web/GameController.java
@@ -1,59 +1,23 @@
 package chess.application.web;
 
+import java.util.Map;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
 
-@Controller
+@RestController
 public class GameController {
     private final GameService gameService;
-    private final JsonTransformer jsonTransformer;
 
     public GameController(GameService gameService) {
-        this.jsonTransformer = new JsonTransformer();
         this.gameService = gameService;
     }
 
-    @GetMapping("/")
-    public String index(Model model) {
-        model.addAllAttributes(gameService.modelReady());
-        return "index";
-    }
-
-    @GetMapping("/start")
-    public String start(Model model) {
-        gameService.start();
-        return play(model);
-    }
-
-    @GetMapping("/load")
-    public String load(Model model) {
-        gameService.load();
-        return play(model);
-    }
-
-    @PostMapping("/move")
-    public String move(Model model, @RequestParam String source, @RequestParam String target) {
-        gameService.move(source, target);
-        if (gameService.isGameFinished()) {
-            return end(model);
-        }
-        return play(model);
-    }
-
-    private String play(Model model) {
-        model.addAllAttributes(gameService.modelPlayingBoard());
-        return "index";
-    }
-
     @GetMapping("/status")
-    public ResponseEntity<String> status() {
-        String statusData = jsonTransformer.render(gameService.modelStatus());
+    public ResponseEntity<Map<String, Object>> status() {
+        Map<String, Object> statusData = gameService.modelStatus();
         return ResponseEntity.ok().body(statusData);
     }
 
@@ -61,12 +25,6 @@ public class GameController {
     public ResponseEntity<Void> save() {
         gameService.save();
         return ResponseEntity.status(HttpStatus.CREATED).build();
-    }
-
-    @GetMapping("/end")
-    public String end(Model model) {
-        model.addAllAttributes(gameService.end());
-        return "result";
     }
 
     @ExceptionHandler
